# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jeigo API",
        "description": "Dictionary API",
        "version": "1.0.0"
    },
    "paths": {
...
        "entries/en": {
            "get": {
                "tags": [
                    "Words"
                ],
                "summary": "List words",
                "operationId": "fd4df9f5555cd5cb411dd39f6262ed2b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "benches"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/entries/en"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "next_cursor": {
                                            "type": "string",
                                            "example": "eyJ3b3Jkcy5pZCI6MzE0ODksIl9wb2ludHNUb05leHRJdGVtcyI6dHJ1ZX0"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/entries/en?cursor=eyJ3b3Jkcy5pZCI6MzE0ODksIl9wb2ludHNUb05leHRJdGVtcyI6dHJ1ZX0"
                                        },
                                        "prev_cursor": {
                                            "type": "string",
                                            "example": "eyJ3b3Jkcy5pZCI6MzE0NzAsIl9wb2ludHNUb05leHRJdGVtcyI6ZmFsc2V9"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/entries/en?cursor=eyJ3b3Jkcy5pZCI6MzE0NzAsIl9wb2ludHNUb05leHRJdGVtcyI6ZmFsc2V9"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "entries/en/{word}": {
            "get": {
                "tags": [
                    "Words"
                ],
                "summary": "Get word details",
                "description": "Returns details about the given word, including meanings, phonetics, and source links.",
                "operationId": "getWordDetails",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "The word to fetch information for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "word": {
                                                "type": "string",
                                                "example": "documentation"
                                            },
                                            "phonetic": {
                                                "type": "string",
                                                "example": "/ˌdɒkjʊmənˈteɪʃən/"
                                            },
                                            "phonetics": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "text": {
                                                            "type": "string",
                                                            "example": "/ˌdɒkjʊmənˈteɪʃən/"
                                                        },
                                                        "audio": {
                                                            "type": "string",
                                                            "example": ""
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "meanings": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "partOfSpeech": {
                                                            "type": "string",
                                                            "example": "noun"
                                                        },
                                                        "definitions": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "definition": {
                                                                        "type": "string",
                                                                        "example": "Something transposed from a thought to a document; the written account of an idea."
                                                                    },
                                                                    "synonyms": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "antonyms": {
                                                                        "type": "array",
                                                                        "items": {
                                                                            "type": "string"
                                                                        }
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "license": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "CC BY-SA 3.0"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://creativecommons.org/licenses/by-sa/3.0"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "sourceUrls": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "https://en.wiktionary.org/wiki/documentation"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Word not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Word not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}